+++ pkg/analysis_server_plugin/lib/src/plugin_server.dart
@@ -69,6 +69,9 @@ class PluginServer {
   /// The recent state of analysis reults, to be cleared on file changes.
   final _recentState = <String, _PluginState>{};
 
+  /// The next modification stamp for a changed file in the [resourceProvider].
+  int _overlayModificationStamp = 0;
+
   PluginServer({
     required ResourceProvider resourceProvider,
     required List<Plugin> plugins,
@@ -77,29 +80,6 @@ class PluginServer {
     for (var plugin in plugins) {
       plugin.register(_registry);
     }
-  }
-
-  /// Handles an 'analysis.setContextRoots' request.
-  Future<protocol.AnalysisSetContextRootsResult> handleAnalysisSetContextRoots(
-      protocol.AnalysisSetContextRootsParams parameters) async {
-    var currentContextCollection = _contextCollection;
-    if (currentContextCollection != null) {
-      _contextCollection = null;
-      await currentContextCollection.dispose();
-    }
-
-    var includedPaths = parameters.roots.map((e) => e.root).toList();
-    var contextCollection = AnalysisContextCollectionImpl(
-      resourceProvider: _resourceProvider,
-      includedPaths: includedPaths,
-      byteStore: _byteStore,
-      sdkPath: _sdkPath,
-      fileContentCache: FileContentCache(_resourceProvider),
-    );
-    _contextCollection = contextCollection;
-    await _analyzeAllFilesInContextCollection(
-        contextCollection: contextCollection);
-    return protocol.AnalysisSetContextRootsResult();
   }
 
   /// Handles an 'edit.getFixes' request.
@@ -193,8 +173,6 @@ class PluginServer {
 
   /// This method is invoked when a new instance of [AnalysisContextCollection]
   /// is created, so the plugin can perform initial analysis of analyzed files.
-  ///
-  /// By default analyzes every [AnalysisContext] with [_analyzeFiles].
   Future<void> _analyzeAllFilesInContextCollection({
     required AnalysisContextCollection contextCollection,
   }) async {
@@ -326,27 +304,36 @@ class PluginServer {
       case protocol.ANALYSIS_REQUEST_GET_NAVIGATION:
       case protocol.ANALYSIS_REQUEST_HANDLE_WATCH_EVENTS:
         result = null;
+
       case protocol.ANALYSIS_REQUEST_SET_CONTEXT_ROOTS:
         var params =
             protocol.AnalysisSetContextRootsParams.fromRequest(request);
-        result = await handleAnalysisSetContextRoots(params);
+        result = await _handleAnalysisSetContextRoots(params);
+
       case protocol.ANALYSIS_REQUEST_SET_PRIORITY_FILES:
       case protocol.ANALYSIS_REQUEST_SET_SUBSCRIPTIONS:
       case protocol.ANALYSIS_REQUEST_UPDATE_CONTENT:
+        var params = protocol.AnalysisUpdateContentParams.fromRequest(request);
+        result = await _handleAnalysisUpdateContent(params);
+
       case protocol.COMPLETION_REQUEST_GET_SUGGESTIONS:
       case protocol.EDIT_REQUEST_GET_ASSISTS:
       case protocol.EDIT_REQUEST_GET_AVAILABLE_REFACTORINGS:
         result = null;
+
       case protocol.EDIT_REQUEST_GET_FIXES:
         var params = protocol.EditGetFixesParams.fromRequest(request);
         result = await handleEditGetFixes(params);
+
       case protocol.EDIT_REQUEST_GET_REFACTORING:
         result = null;
+
       case protocol.PLUGIN_REQUEST_SHUTDOWN:
         _channel.sendResponse(protocol.PluginShutdownResult()
             .toResponse(request.id, requestTime));
         _channel.close();
         return null;
+
       case protocol.PLUGIN_REQUEST_VERSION_CHECK:
         var params = protocol.PluginVersionCheckParams.fromRequest(request);
         result = await handlePluginVersionCheck(params);
@@ -356,6 +343,119 @@ class PluginServer {
           error: RequestErrorFactory.unknownRequest(request.method));
     }
     return result.toResponse(request.id, requestTime);
+  }
+
+  /// Handles files that might have been affected by a content change of
+  /// one or more files. The implementation may check if these files should
+  /// be analyzed, do such analysis, and send diagnostics.
+  ///
+  /// By default invokes [_analyzeFiles] only for files that are analyzed in
+  /// this [analysisContext].
+  Future<void> _handleAffectedFiles({
+    required AnalysisContext analysisContext,
+    required List<String> paths,
+  }) async {
+    var analyzedPaths = paths
+        .where(analysisContext.contextRoot.isAnalyzed)
+        .toList(growable: false);
+
+    await _analyzeFiles(
+      analysisContext: analysisContext,
+      paths: analyzedPaths,
+    );
+  }
+
+  /// Handles an 'analysis.setContextRoots' request.
+  Future<protocol.AnalysisSetContextRootsResult> _handleAnalysisSetContextRoots(
+      protocol.AnalysisSetContextRootsParams parameters) async {
+    var currentContextCollection = _contextCollection;
+    if (currentContextCollection != null) {
+      _contextCollection = null;
+      await currentContextCollection.dispose();
+    }
+
+    var includedPaths = parameters.roots.map((e) => e.root).toList();
+    var contextCollection = AnalysisContextCollectionImpl(
+      resourceProvider: _resourceProvider,
+      includedPaths: includedPaths,
+      byteStore: _byteStore,
+      sdkPath: _sdkPath,
+      fileContentCache: FileContentCache(_resourceProvider),
+    );
+    _contextCollection = contextCollection;
+    await _analyzeAllFilesInContextCollection(
+        contextCollection: contextCollection);
+    return protocol.AnalysisSetContextRootsResult();
+  }
+
+  /// Handles an 'analysis.updateContent' request.
+  ///
+  /// Throws a [RequestFailure] if the request could not be handled.
+  Future<protocol.AnalysisUpdateContentResult> _handleAnalysisUpdateContent(
+      protocol.AnalysisUpdateContentParams parameters) async {
+    var changedPaths = <String>{};
+    var paths = parameters.files;
+    paths.forEach((String path, Object overlay) {
+      // Prepare the old overlay contents.
+      String? oldContent;
+      try {
+        if (_resourceProvider.hasOverlay(path)) {
+          oldContent = _resourceProvider.getFile(path).readAsStringSync();
+        }
+      } catch (_) {
+        // Leave `oldContent` empty.
+      }
+
+      // Prepare the new contents.
+      String? newContent;
+      if (overlay is protocol.AddContentOverlay) {
+        newContent = overlay.content;
+      } else if (overlay is protocol.ChangeContentOverlay) {
+        if (oldContent == null) {
+          // The server should only send a ChangeContentOverlay if there is
+          // already an existing overlay for the source.
+          throw RequestFailure(
+              RequestErrorFactory.invalidOverlayChangeNoContent());
+        }
+        try {
+          newContent =
+              protocol.SourceEdit.applySequence(oldContent, overlay.edits);
+        } on RangeError {
+          throw RequestFailure(
+              RequestErrorFactory.invalidOverlayChangeInvalidEdit());
+        }
+      } else if (overlay is protocol.RemoveContentOverlay) {
+        newContent = null;
+      }
+
+      if (newContent != null) {
+        _resourceProvider.setOverlay(
+          path,
+          content: newContent,
+          modificationStamp: _overlayModificationStamp++,
+        );
+      } else {
+        _resourceProvider.removeOverlay(path);
+      }
+
+      changedPaths.add(path);
+    });
+    await _handleContentChanged(changedPaths.toList());
+    return protocol.AnalysisUpdateContentResult();
+  }
+
+  /// Handles the fact that files with [paths] were changed.
+  Future<void> _handleContentChanged(List<String> paths) async {
+    if (_contextCollection case var contextCollection?) {
+      await _forAnalysisContexts(contextCollection, (analysisContext) async {
+        for (var path in paths) {
+          analysisContext.changeFile(path);
+        }
+        var affected = await analysisContext.applyPendingFileChanges();
+        await _handleAffectedFiles(
+            analysisContext: analysisContext, paths: affected);
+      });
+    }
   }
 
   Future<void> _handleRequest(Request request) async {