+++ pkg/analyzer/lib/src/dart/resolver/extension_member_resolver.dart
@@ -84,7 +84,7 @@ class ExtensionMemberResolver {
   /// returns [ResolutionResult.ambiguous].
   ResolutionResult findExtension(
       DartType type, SyntacticEntity nameEntity, Name name) {
-    var extensions = _resolver.libraryFragment.scope.accessibleExtensions
+    var extensions = _resolver.libraryFragment.accessibleExtensions
         .havingMemberWithBaseName(name)
         .applicableTo(
           targetLibrary: _resolver.definingLibrary,