+++ pkg/dev_compiler/lib/js/ddc/ddc_module_loader.js
@@ -360,6 +360,14 @@ if (!self.dart_library) {
     // Map from module name to corresponding app to proxy library map.
     let _proxyLibs = new Map();
 
+    /**
+     * Returns an instantiated module given its module name.
+     *
+     * Note: this method is not meant to be used outside DDC generated code,
+     * however it is currently being used in many places becase DDC lacks an
+     * Embedding API. This API will be removed in the future once the Embedding
+     * API is established.
+     */
     function import_(name, appName) {
       // For backward compatibility.
       if (!appName && _lastStartedSubapp) {
@@ -609,8 +617,25 @@ if (!self.dart_library) {
     let _firstStartedAppName;
     let _lastStartedSubapp;
 
-    /// Starts a subapp that is identified with `uuid`, `moduleName`, and
-    /// `libraryName` inside a parent app that is identified by `appName`.
+    /**
+     * Runs a Dart program's main method.
+     *
+     * Intended to be invoked by the bootstrapping code where the application
+     * is being embedded.
+     *
+     * Because multiple programs can be part of a single application on a page
+     * at once, we identify the entrypoint using multiple keys: `appName`
+     * denotes the parent application, this program within that application
+     * (aka. the subapp) is identified by an `uuid`. Finally the entrypoint
+     * method is located from the `moduleName` and `libraryName`.
+     *
+     * Often, when a page contains a single app, the uuid is a fixed trivial
+     * value like '00000000-0000-0000-0000-000000000000'.
+     *
+     * This is one of the current public Embedding APIs exposed by DDC.
+     * Note: this API will be replaced by `dartDevEmbedder.runMain` in the
+     * future.
+     */
     function start(appName, uuid, moduleName, libraryName, isReload) {
       console.info(
         `DDC: Subapp Module [${appName}:${moduleName}:${uuid}] is starting`);
@@ -676,6 +701,75 @@ if (!self.dart_library) {
       library.main([]);
     }
     dart_library.start = start;
+
+
+    /**
+     * Configure the DDC runtime.
+     *
+     * Must be called before invoking [start].
+     *
+     * The configuration parameter is an object that may provide any of the
+     * following keys:
+     *
+     *   - weakNullSafetyErrors: throw errors when types violate sound null
+     *        safety (deprecated).
+     *
+     *   - nonNullAsserts: insert non-null assertions no non-nullable method
+     *        parameters (deprecated, was used to aid in null safety
+     *        migrations).
+     *
+     *   - nativeNonNullAsserts: inject non-null assertions checks to validate
+     *        that browser APIs are sound. This is helpful because browser APIs
+     *        are generated from IDLs and cannot be proven to be correct
+     *        statically.
+     *
+     *   - jsInteropNonNullAsserts: inject non-null assertiosn to check
+     *        nullability of `package:js` JS-interop APIs. The new
+     *        `dart:js_interop` always includes non-null assertions.
+     *
+     *   - dynamicModuleLoader: provide the implementation of dynamic module
+     *        loading. Dynamic modules is an experimental feature.
+     *        The DDC runtime delegates to the embedder how code for dynamic
+     *        modules is downloaded. This entry in the configuration must be a
+     *        function with the signature:
+     *            `function(String, onLoad)`
+     *        It accepts a `String` containing a Uri that denotes the module to
+     *        be loaded. When the load completes, the loader must invoke
+     *        `onLoad` and provide the module name of the dynamic module to it.
+     *
+     *        Note: eventually we may want to make the loader return a promise.
+     *        We avoided that for now because it interfereres with our testing
+     *        in d8.
+     */
+    dart_library.configure = function configure(appName, configuration) {
+      let runtimeLibrary = dart_library.import("dart_sdk", appName).dart;
+      if (!!configuration.weakNullSafetyErrors) {
+        runtimeLibrary.weakNullSafetyErrors(configuration.weakNullSafetyErrors);
+      }
+      if (!!configuration.nonNullAsserts) {
+        runtimeLibrary.nonNullAsserts(configuration.nonNullAsserts);
+      }
+      if (!!configuration.nativeNonNullAsserts) {
+        runtimeLibrary.nativeNonNullAsserts(configuration.nativeNonNullAsserts);
+      }
+      if (!!configuration.jsInteropNonNullAsserts) {
+        runtimeLibrary.jsInteropNonNullAsserts(
+           configuration.jsInteropNonNullAsserts);
+      }
+      if (!!configuration.dynamicModuleLoader) {
+        let loader = configuration.dynamicModuleLoader;
+        runtimeLibrary.setDynamicModuleLoader(loader, (moduleName) => {
+          //  As mentioned in the docs above, loader will not invoke the
+          //  entrypoint, but just return the moduleName to find where to call
+          //  it.  By using the module name, we don't need to expose the
+          //  `import` as part of the embedding API.
+          //  In the future module system, this will change to a library name,
+          //  rather than a module name.
+          return import_(moduleName, appName).__dynamic_module_entrypoint__();
+        });
+      }
+    }
+
   })(dart_library);
 }
 
@@ -757,14 +851,20 @@ if (!self.dart_library) {
     };
   }
 
-  // Loads a single script onto the page.
-  // TODO(markzipan): Is there a cleaner way to integrate this?
-  self.$dartLoader.forceLoadScript = function (jsFile) {
+  /**
+   * Loads [jsFile] onto this instance's DDC app's page, then invokes
+   * [onLoad].
+   * @param {string} jsFile
+   * @param {?function()} onLoad Callback after a successful load
+   */
+  self.$dartLoader.forceLoadScript = function (jsFile, onLoad) {
+    // A head element won't be created for D8, so just load synchronously.
     if (self.dart_library.isD8) {
       self.load(jsFile);
+      onLoad?.();
       return;
     }
-    let script = self.dart_library.createScript();
+    let script = dart_library.createScript();
     let policy = {
       createScriptURL: function (src) { return src; }
     };
@@ -772,7 +872,10 @@ if (!self.dart_library) {
       policy = self.trustedTypes.createPolicy('dartDdcModuleUrl', policy);
     }
     script.setAttribute('src', policy.createScriptURL(jsFile));
-    document.head.appendChild(script);
+    script.async = false;
+    script.defer = true;
+    script.onload = onLoad;
+    self.document.head.appendChild(script);
   };
 
   self.$dartLoader.forceLoadModule = function (moduleName) {
@@ -811,14 +914,6 @@ if (!self.dart_library) {
         this.loadConfig.loadScriptFn(this);
       };
 
-      // The current hot restart generation.
-      //
-      // 0-indexed and increases by 1 on every successful hot restart.
-      // This value is read to determine the 'current' hot restart generation
-      // in our hot restart tests. This closely tracks but is not the same as
-      // `hotRestartIteration` in DDC's runtime.
-      this.hotRestartGeneration = 0;
-
       // The current 'intended' hot restart generation.
       //
       // 0-indexed and increases by 1 on every successful hot restart.
@@ -827,11 +922,6 @@ if (!self.dart_library) {
       // This is used to synchronize D8 timers and lookup files to load in
       // each generation for hot restart testing.
       this.intendedHotRestartGeneration = 0;
-
-      // The current hot reload generation.
-      //
-      // 0-indexed and increases by 1 on every successful hot reload.
-      this.hotReloadGeneration = 0;
     }
 
     // True if we are still processing scripts from the script queue.
@@ -1052,12 +1142,6 @@ if (!self.dart_library) {
       }
       this.processAfterLoadOrErrorEvent();
     };
-
-    // Initiates a hot reload.
-    // TODO(markzipan): This function is currently stubbed out for testing.
-    hotReload() {
-      this.hotReloadGeneration += 1;
-    }
   };
 
   let policy = {
@@ -1280,6 +1364,14 @@ if (!self.deferred_loader) {
     // These are the result of calling a library's initialization function.
     libraries = Object.create(null);
 
+    pendingHotReloadLibraryNames = null;
+    pendingHotReloadFileUrls = null;
+    pendingHotReloadLibraryInitializers = Object.create(null);
+
+    // The name of the entrypoint module. Set when the application starts for
+    // the first time and used during a hot restart.
+    savedEntryPointLibraryName = null;
+
     // The current hot restart generation.
     //
     // 0-indexed and increases by 1 on every successful hot restart.
@@ -1293,6 +1385,11 @@ if (!self.deferred_loader) {
     // TODO(nshahan): Set to true at the start of the hot reload process.
     hotReloadInProgress = false;
 
+    // The current hot reload generation.
+    //
+    // 0-indexed and increases by 1 on every successful hot reload.
+    hotReloadGeneration = 0;
+
     // The name of the entrypoint module. Set when the application starts for
     // the first time and used during a hot restart.
     savedEntryPointLibraryName = null;
@@ -1303,10 +1400,12 @@ if (!self.deferred_loader) {
 
     // See docs on `DartDevEmbedder.runMain`.
     defineLibrary(libraryName, initializer) {
-      if (this.hotReloadInProgress) {
-        // TODO(nshahan): Store initialization functions on the side once a
-        // hot reload starts because the app could continue to run until the
-        // newly compiled libraries all arrive in the browser.
+      // TODO(nshahan): Make this test stronger and check for generations. A
+      // library that is part of a pending hot reload could also be defined as
+      // part of the previous generation.
+      if (this.hotReloadInProgress
+        && this.pendingHotReloadLibraryNames.includes(libraryName)) {
+        this.pendingHotReloadLibraryInitializers[libraryName] = initializer;
       } else {
         this.libraryInitializers[libraryName] = initializer;
       }
@@ -1403,17 +1502,62 @@ if (!self.deferred_loader) {
 
     /**
      * Begins a hot reload operation.
+     *
+     * @param {Array<String>} filesToLoad The urls of the files that contain
+     * the libraries to hot reload.
+     * @param {Array<String>} librariesToReload The names of the libraries to
+     * hot reload.
      */
     async hotReloadStart(filesToLoad, librariesToReload) {
-      // TODO(nshahan): Request newly compiled files and call `hotReloadEnd()`
-      // when complete.
+      this.hotReloadInProgress = true;
+      this.pendingHotReloadFileUrls ??= filesToLoad;
+      this.pendingHotReloadLibraryNames ??= librariesToReload;
+      // Trigger download of the new library versions.
+      let reloadFilePromises = [];
+      for (let file of this.pendingHotReloadFileUrls) {
+        reloadFilePromises.push(
+          new Promise((resolve) => {
+            self.$dartLoader.forceLoadScript(file, resolve)
+          })
+        );
+      }
+      await Promise.all(reloadFilePromises).then((_) => {
+        this.hotReloadEnd();
+      });
     }
 
     /**
      * Completes a hot reload operation.
      */
-    hotReloadEnd(librariesToReload) {
-      // TODO(nshahan): Initialize and link all the newly compiled libraries.
+    hotReloadEnd() {
+      let oldLibraries = Object.create(null);
+      // Remove all the current library values to ensure all reloaded libraries
+      // get fresh copies of each other when imported.
+      for (let name of this.pendingHotReloadLibraryNames) {
+        oldLibraries[name] = this.libraries[name];
+        this.libraries[name] = null;
+      }
+      // Initialize all reloaded libraries.
+      for (let name of this.pendingHotReloadLibraryNames) {
+        let currentLibrary = oldLibraries[name];
+        if (currentLibrary == null) {
+          currentLibrary = this.createEmptyLibrary();
+        }
+        let initializer = this.pendingHotReloadLibraryInitializers[name];
+        this.libraryInitializers[name] = initializer;
+        this.libraries[name] = initializer(currentLibrary);
+      }
+      // Link all reloaded libraries.
+      for (let name in this.pendingHotReloadLibraryInitializers) {
+        let currentLibrary = this.libraries[name];
+        currentLibrary.link();
+      }
+      // Cleanup.
+      this.pendingHotReloadLibraryInitializers = Object.create(null);
+      this.pendingHotReloadLibraryNames = null;
+      this.pendingHotReloadFileUrls = null;
+      this.hotReloadInProgress = false;
+      this.hotReloadGeneration += 1;
     }
 
     /**
@@ -1497,6 +1641,21 @@ if (!self.deferred_loader) {
     }
 
     /**
+     * DDC's entrypoint for triggering a hot reload.
+     * 
+     * Previous generations may continue to run until all specified files
+     * have been loaded and initialized.
+     *
+     * @param {Array<String>} filesToLoad The urls of the files that contain
+     * the libraries to hot reload.
+     * @param {Array<String>} librariesToReload The names of the libraries to
+     * hot reload.
+     */
+    async hotReload(filesToLoad, librariesToReload) {
+      await libraryManager.hotReloadStart(filesToLoad, librariesToReload);
+    }
+
+    /**
      * Immediately triggers a hot restart of the application losing all state
      * and running the main method again.
      */
@@ -1504,6 +1663,15 @@ if (!self.deferred_loader) {
       self.$dartReloadModifiedModules(
         libraryManager.savedEntryPointLibraryName,
         () => { libraryManager.hotRestart(); });
+    }
+
+
+    /**
+     * @return {Number} The current hot reload generation of the running
+     *  application.
+     */
+    get hotReloadGeneration() {
+      return libraryManager.hotReloadGeneration;
     }
 
     /**