+++ pkg/reload_test/lib/ddc_helpers.dart
@@ -122,6 +122,8 @@ self.\$dartLoader.loader = loader;
 // Append hot reload runner-specific logic.
 let modifiedFilesPerGeneration = ${_encoder.convert(modifiedFilesPerGeneration)};
 let previousGenerations = new Set();
+
+// Append a helper function for hot restart.
 self.\$dartReloadModifiedModules = function(subAppName, callback) {
   let expectedName = "$entrypointModuleName";
   if (subAppName !== expectedName) {
@@ -152,6 +154,30 @@ self.\$dartReloadModifiedModules = function(subAppName, callback) {
 
   // Run main.
   callback();
+}
+
+// Append a helper function for hot reload.
+self.\$injectedFilesAndLibrariesToReload = function() {
+  // Resolve the next generation's directory and load all modified files.
+  let nextGeneration = self.dartDevEmbedder.hotReloadGeneration + 1;
+  if (previousGenerations.has(nextGeneration)) {
+    throw Error('Fatal error: Previous generations are being re-run.');
+  }
+  previousGenerations.add(nextGeneration);
+  let modifiedFilePaths = modifiedFilesPerGeneration[nextGeneration];
+  // Stop if the next generation does not exist.
+  if (modifiedFilePaths == void 0) {
+    return;
+  }
+  let fileUrls = [];
+  let libraryIds = [];
+  for (let i = 0; i < modifiedFilePaths.length; i++) {
+    let modifiedFileId =  modifiedFilePaths[i][0];
+    let modifiedFilePath = modifiedFilePaths[i][1];
+    libraryIds.push(modifiedFileId);
+    fileUrls.push(modifiedFilePath);
+  }
+  return [fileUrls, libraryIds];
 }
 
 // D8 does not support the core Timer API methods beside `setTimeout` so our
@@ -417,6 +443,30 @@ let _scriptUrls = {
     // Append hot reload runner-specific logic.
     let modifiedFilesPerGeneration = ${_encoder.convert(modifiedFilesPerGeneration)};
     let previousGenerations = new Set();
+
+    self.\$injectedFilesAndLibrariesToReload = function() {
+      // Resolve the next generation's directory and load all modified files.
+      let nextGeneration = self.dartDevEmbedder.hotReloadGeneration + 1;
+      if (previousGenerations.has(nextGeneration)) {
+        throw Error('Fatal error: Previous generations are being re-run.');
+      }
+      previousGenerations.add(nextGeneration);
+      let modifiedFilePaths = modifiedFilesPerGeneration[nextGeneration];
+      // Stop if the next generation does not exist.
+      if (modifiedFilePaths == void 0) {
+        return;
+      }
+      let fileUrls = [];
+      let libraryIds = [];
+      for (let i = 0; i < modifiedFilePaths.length; i++) {
+        let modifiedFileId =  modifiedFilePaths[i][0];
+        let modifiedFilePath = modifiedFilePaths[i][1];
+        libraryIds.push(modifiedFileId);
+        fileUrls.push(modifiedFilePath);
+      }
+      return [fileUrls, libraryIds];
+    }
+
     self.\$dartReloadModifiedModules = function(subAppName, callback) {
       let expectedName = "$entrypointModuleName";
       if (subAppName !== expectedName) {