+++ sdk/lib/_internal/js_dev_runtime/patch/internal_patch.dart
@@ -4,6 +4,7 @@
 
 import 'dart:core' hide Symbol;
 import 'dart:core' as core show Symbol;
+import 'dart:async' show Completer;
 import 'dart:_js_primitives' show printString;
 import 'dart:_internal' show patch;
 import 'dart:_interceptors' show JSArray;
@@ -67,5 +68,32 @@ bool isSentinel(dynamic value) => throw UnsupportedError('isSentinel');
 T unsafeCast<T>(dynamic v) => v;
 
 @patch
-Future<Object?> loadDynamicModule({Uri? uri, Uint8List? bytes}) =>
-    throw 'Unsupported operation';
+Future<Object?> loadDynamicModule({Uri? uri, Uint8List? bytes}) {
+  if (bytes != null) {
+    throw ArgumentError('DDC implementation of dynamic modules doesn\'t'
+        ' accept bytes as input');
+  }
+  if (uri == null) {
+    throw ArgumentError('DDC implementation of dynamic modules expects a'
+        'non-null Uri input.');
+  }
+  if (dart.dynamicModuleLoader == null) {
+    throw StateError('Dynamic module loader has not be configured.');
+  }
+  var completer = Completer<Object?>();
+  void _callback(String moduleName) {
+    try {
+      var result = JS('!', '#(#)', dart.dynamicEntrypointHelper, moduleName);
+      completer.complete(result);
+    } catch (e, st) {
+      completer.completeError(e, st);
+    }
+  }
+
+  try {
+    JS('!', '#(#, #)', dart.dynamicModuleLoader, uri.toString(), _callback);
+  } catch (e, st) {
+    completer.completeError(e, st);
+  }
+  return completer.future;
+}